Changed NodeInfo
There were two fields, .iex and .iin that referenced the external and internal sides of the node. This was replaced by .side whose type is parametrized. For now, we will work with Tuples of integers that refer to the pressure tap of the external face and pressure tap of the internal node. In th future we can implement other things.